Simplify command dispatching examples
The snippets describing command dispatching are a little to beefy. Trim
down the sample to contain no import statements and class signatures,
and pairing example functions were applicable.
Additionally, remove the CommandCallback example for the CommandGateway,
 as this use case can simply be derived by users themselves.
Lastly, mentions of 'Note' should be replaced by a short header
describing what the 'note' is about

#45