Revamp Command Model Testing section
The Command Model testing section more so resembled an Axon 2
description, with some updates which are in line with Axon 3 and even
fewer matching Axon 4. It is thus beneficial to realign this entire
chapter, by updating it to the latest state of the Aggregate TestFixture
 suite. In the process, update the indentation for clarity, fix typos
 when applicable and update the notice section to have an actual title.
Lastly, introduce a section about validating deadlines, as that's
currently completely missing in the reference guide.

#45