Revamp the Distributing the Command Bus section
Rename this section to "The Command Bus - Distributed", providing a
short introduction on what a distributed command bus means.
Follow this up with a section explaining what the Local Segment is, as
this currently does not have a dedicated section. Next, a Load Factor
section can be provided, removing the notice at the end on the
`DistributedCommandBus` section. Then, a distinct Routing Strategy
should be provided, retrieving the current portion in the
`DistributedCommandBus` and expanding on it greatly with more samples on
 how to build and how the configure. Lastly, draft the
 AxonServerCommandBus section based on what was present, enhanced with
 some additional configuration specifics.

#188