Minor adjustments
Adjust the introduction of the JGroups extension slightly, adding a
link to JGroups and adjusting readability, with an added reference to
the spring boot starter. Also adjust the configuring with spring
section, as it in essence is a combination of configuring with and
without Spring. This should be apparent from the documentation and
sample. Lastly, include the enable and load factor properties to the
sample properties file

#188