Replace the Configuring processors section
The Configuring processors section was a broad combination of several
configurable options, focusing on Subscribing and Tracking specifics.
Added, it is a non-exhaustive list, which also contains Saga specifics.
The Saga section should be removed entirely and placed in the Saga
description. Furthermore, the section should be renamed to describe
generic Event Processor configuration. Lastly, this section should carry
 the remainder of configurable options as well.

#189