Adjust the page around serialization
Change the page about serialization to no longer be specific about
events. Back in the day this was reasonable, but serializers are used
throughout, for a lot of objects within Axon. Hence, it should not be a
part of the events subsection, but rather have a presence on the main
page. Furthermore, we should take the opportunity to make some
adjustments in the serialization.md page, like making it less event and
XStream focussed. Lastly a warning around XStream in combination with
JDK17, which doesn't always work, is good to include.

#enhancement/serialization