Revamp the Handler Enhancer Definition section
The Handler Enhancers section contains some minor inconsistency that
should be resolved. Like, usage of the `annotationAttributes` method,
which is deprecated. Furthermore, the configuration section shouldn't
point to HandlerDefinition configuration at all, as that should be a
separate section. Lasty, we can add the new
registerHandlerEnhancerDefinition method of the Configurer, following PR
 https://github.com/AxonFramework/AxonFramework/pull/2545

#enhancement/af-4.7-adjustments